Make some external blank cells transparent
For all blank cells outside the entity body that are between
non-whitespace, make them explicitly transparent.  This is most
noticeable in the sword fish and the submarine.

In shark entity, remove
  default_color => 'WHITE'
because it's followed by
  default_color => 'CYAN'

In fishhook entity, remove
  trans_char    => ' '
because now the transparent cells are specified in the image.  Besides,
the correct property name is "transparent", not "trans_char".

Fix missing cell color in big_fish_mask.